fix(engine): removing globalThis polyfill
