+++ packages/@lwc/engine/src/framework/main.ts
@@ -6,7 +6,6 @@
  */
 
 // Polyfills
-import '../polyfills/global-this/main';
 import '../polyfills/proxy-concat/main';
 import '../polyfills/aria-properties/main';
