--- packages/@lwc/engine/src/polyfills/global-this/detect.ts
@@ -1,9 +1,0 @@
-/*
- * Copyright (c) 2018, salesforce.com, inc.
- * All rights reserved.
- * SPDX-License-Identifier: MIT
- * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
- */
-export default function detect(): boolean {
-    return typeof globalThis !== 'object';
-}
