--- packages/@lwc/engine/src/polyfills/global-this/main.ts
@@ -1,12 +1,0 @@
-/*
- * Copyright (c) 2018, salesforce.com, inc.
- * All rights reserved.
- * SPDX-License-Identifier: MIT
- * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
- */
-import detect from './detect';
-import apply from './polyfill';
-
-if (detect()) {
-    apply();
-}
