--- packages/@lwc/engine/src/polyfills/global-this/polyfill.js
@@ -1,17 +1,0 @@
-/*
- * Copyright (c) 2018, salesforce.com, inc.
- * All rights reserved.
- * SPDX-License-Identifier: MIT
- * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
- */
-export default function apply() {
-    Object.defineProperty(Object.prototype, '__magic__', {
-        get: function() {
-            return this;
-        },
-        configurable: true,
-    });
-    // eslint-disable-next-line no-undef
-    __magic__.globalThis = __magic__;
-    delete Object.prototype.__magic__;
-}
