+++ packages/@lwc/engine/test.html
@@ -1,0 +1,107 @@
+<html>
+    <head>
+        <script src="dist/umd/es2017/engine.js"></script>
+        <script>
+        var { createElement, LightningElement, registerComponent, registerTemplate, register } = Engine;
+        </script>
+    </head>
+    <body>
+        <script>
+            // const fooHTML = registerTemplate(function($api, $cmp) {
+            //     debugger;
+            //     return [$api.c('x-bar', Bar, { key: 1 }, [])];
+            // });
+            // class Foo extends LightningElement {
+
+            // }
+            // registerComponent(Foo, { name: 'Foo', tmpl: fooHTML });
+            // const barHTML = registerTemplate(function($api, $cmp) {
+            //     debugger;
+            //     return [$api.h('p', { key: 1 }, [])];
+            // });
+            // class Bar extends LightningElement {
+
+            // }
+            // registerComponent(Bar, { name: 'Bar', tmpl: barHTML });
+            // const elm = createElement('x-foo', { is: Foo }); 
+            // document.body.appendChild(elm);
+
+            const calls = [];
+
+            class Child extends LightningElement {
+                constructor() {
+                    super();
+                    calls.push('child:constructor');
+                    console.error('child:constructor');
+                }
+                connectedCallback() {
+                    calls.push('child:connectedCallback');
+                }
+                render() {
+                    calls.push('child:render');
+                    return emptyTemplate;
+                }
+                renderedCallback() {
+                    calls.push('child:renderedCallback');
+                }
+                disconnectedCallback() {
+                    calls.push('child:disconnectedCallback');
+                }
+            }
+            const emptyTemplate = registerTemplate(function () { return [] });
+            registerComponent(Child, {});
+
+            class Root extends LightningElement {
+                constructor() {
+                    super();
+                    this.state = {
+                        show: false,
+                    };
+                }
+                show() {
+                    this.state.show = true;
+                }
+                hide() {
+                    this.state.show = false;
+                }
+                render() {
+                    calls.push('root:render');
+                    return html;
+                }
+                renderedCallback() {
+                    calls.push('root:renderedCallback');
+                }
+            }
+            const html = registerTemplate(function($api, $cmp) {
+                debugger;
+                return [$api.h('div', { key: 2 }, []), $cmp.state.show ? $api.c('x-child', Child, { key: 1 }, []) : null];
+            });
+            Root.publicMethods = ['show', 'hide'];
+            Root.track = { state: 1 };
+            registerComponent(Root, {});
+
+            const elm = createElement('x-root', { is: Root });
+            document.body.appendChild(elm);
+
+            calls.length = 0;
+            elm.show();
+
+            Promise.resolve().then(() => {
+                elm.hide();
+                Promise.resolve().then(() => {
+                    console.log(calls, [
+                        'root:render',
+                        'child:constructor',
+                        'child:connectedCallback',
+                        'child:render',
+                        'child:renderedCallback',
+                        'root:renderedCallback',
+                        'root:render',
+                        'child:disconnectedCallback',
+                        'root:renderedCallback',
+                    ]);
+                });
+            });
+        </script>
+    </body>
+</html>
