+++ packages/@lwc/features/package.json
@@ -2,14 +2,15 @@
     "name": "@lwc/features",
     "private": true,
     "version": "1.0.2",
-    "description": "LWC Features",
-    "main": "dist/commonjs/flags.js",
-    "module": "dist/esm/flags.js",
+    "description": "LWC Features Flags",
+    "main": "dist/flags.cjs.js",
+    "module": "dist/flags.js",
+    "typings": "types/flags.d.ts",
     "license": "MIT",
     "scripts": {
-        "build": "tsc && tsc --project tsconfig.commonjs.json",
-        "clean": "rm -rf dist/",
-        "test": "jest"
+        "clean": "rm -rf dist/ types/",
+        "test": "jest",
+        "build": "tsc --emitDeclarationOnly && rollup --config ./scripts/rollup/rollup.config.js"
     },
     "files": [
         "dist/"
