--- packages/@lwc/features/tsconfig.commonjs.json
@@ -1,7 +1,0 @@
-{
-    "extends": "./tsconfig.json",
-    "compilerOptions": {
-        "module": "commonjs",
-        "outDir": "dist/commonjs/"
-    }
-}
